add new config option "delete_from_device"
if "delete_from_device" is set to true runs will
be deleted from device, after data downloaded successfully.

+++ antd/main.py
@@ -98,7 +98,7 @@ def downloader():
                         # download runs
                         runs = dev.get_runs()
                         antd.garmin.dump(file, runs)
-                        dev.delete_runs()
+                        if antd.cfg.get_delete_from_device(): dev.delete_runs()
                     _log.info("Closing session.")
                     host.disconnect()
                     _log.info("Excuting plugins.")
